Revert "Page after login and authentication. Adding redux in solution."
This reverts commit a4fcbcf3728aebe716810dc9294c15a248b57d57.
