+++ .gitignore
@@ -7,5 +7,3 @@
 /.vs/VisitorsTracker/DesignTimeBuild/.dtbcache.v2
 /.vs/VisitorsTracker/v16/.suo
 /VisitorsTracker.Core/obj/Debug/netcoreapp3.1/VisitorsTracker.Core.csprojAssemblyReference.cache
-/.vs/VisitorsTracker/DesignTimeBuild/.dtbcache.v2
-/.vs/VisitorsTracker/v16/.suo
