+++ VisitorsTracker/ClientApp/package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
     "bootstrap": "^4.1.3",
     "jquery": "^3.5.1",
     "merge": "^1.2.1",
