+++ VisitorsTracker/ClientApp/src/App.test.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { MemoryRouter } from 'react-router-dom';
+import App from './App';
+
+it('renders without crashing', async () => {
+  const div = document.createElement('div');
+  ReactDOM.render(
+    <MemoryRouter>
+      <App />
+    </MemoryRouter>, div);
+  await new Promise(resolve => setTimeout(resolve, 1000));
+});
