+++ VisitorsTracker/ClientApp/src/components/Layout/Layout.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { Container } from 'reactstrap';
-import LeftSidebarWrapper from '../left-sidebar/index';
 import { NavMenu } from '../NavMenu/NavMenu';
 
 export class Layout extends Component {
@@ -10,7 +9,6 @@ export class Layout extends Component {
     return (
       <div>
         <NavMenu />
-        <LeftSidebarWrapper />
         <Container>
           {this.props.children}
         </Container>
