--- VisitorsTracker/ClientApp/src/components/avatar/index.js
@@ -1,3 +1,0 @@
-import CustomAvatar from './custom-avatar'
-
-export default CustomAvatar;
