--- VisitorsTracker/ClientApp/src/components/header-profile/header-profile.css
@@ -1,31 +1,0 @@
-.header-profile-root {
-    flex-grow: 1;
-    width: 100%;    
-    display: flex;
-    justify-content: center;
-}
-
-.menuButton {
-    margin-right: 2em;
-}
-
-.title {
-    flex-grow: 1;
-}
-
-.MiddleAvatar {
-    margin: 10px !important;
-    width: 100px !important;
-    height: 100px !important;
-}
-.bigAvatar {
-    margin: 10px !important;
-    width: 150px !important;
-    height: 150px !important;
-}
-
-.littleAvatar {
-    margin: 10px !important;
-    width: 75px !important;
-    height: 75px !important;
-}
