--- VisitorsTracker/ClientApp/src/components/header-profile/index.js
@@ -1,3 +1,0 @@
-import HeaderProfile from './header-profile';
-
-export default HeaderProfile;
