--- VisitorsTracker/ClientApp/src/components/left-sidebar/index.js
@@ -1,3 +1,0 @@
-import LeftSidebar from './left-sidebar';
-
-export default LeftSidebar;
