+++ VisitorsTracker/ClientApp/src/containers/logout.js
@@ -1,7 +1,5 @@
 import React, { Component } from 'react';
-import { connect } from "react-redux";
 import { withRouter, Redirect } from "react-router-dom";
-import { logout } from "../actions/authActions";
 
 class Logout extends Component {
     componentWillMount(){
@@ -9,7 +7,7 @@ class Logout extends Component {
     }
 
     componentDidMount() {
-        //this.props.logout();
+        // this.props.logout();
     }
 
     render(){
@@ -21,18 +19,5 @@ class Logout extends Component {
     }
 };
 
-const mapStateToProps = (state) => {
-    return {
-      login: state.login
-    };
-  };
   
-  const mapDispatchToProps = (dispatch) => {
-    return {
-      logout: () => {
-        dispatch(logout());
-      }
-    }
-  };
-  
-  export default withRouter(connect(mapStateToProps, mapDispatchToProps)(Logout));
+  export default Logout;
