+++ VisitorsTracker/ClientApp/src/services/authenticationService.js
@@ -5,8 +5,6 @@ const baseService = new VisitorsTrackerService();
 export default class AuthenticationService {
 
     auth = async (data) => {
-        console.log("auth service method");
-        console.log(data);
         const res = await baseService.setResource(`Authentication/verify/${data.userId}/${data.token}`);
         return !res.ok
             ? { error: await res.text() }
@@ -14,7 +12,7 @@ export default class AuthenticationService {
     }
 
     setGoogleLogin = async (data) => {
-        console.log("setGoogleLogin service method");
+        console.log("setGoogleLogin");
         console.log(data);
         const res = await baseService.setResource('Authentication/GoogleLogin', data);
         return !res.ok
