--- VisitorsTracker/ClientApp/src/store/configureStore.js
@@ -1,29 +1,0 @@
-import { applyMiddleware, combineReducers, compose, createStore } from 'redux';
-import thunk from 'redux-thunk';
-import { routerMiddleware } from 'react-router-redux';
-import rootReducers from '../reducers/root';
-
-export default function configureStore(history, initialState) {
-
-  const middleware = [
-    thunk,
-    routerMiddleware(history)
-  ];
-
-  // In development, use the browser's Redux dev tools extension if installed
-  const enhancers = [];
-  const isDevelopment = process.env.NODE_ENV === 'development';
-  if (isDevelopment && typeof window !== 'undefined' && window.devToolsExtension) {
-    enhancers.push(window.devToolsExtension());
-    }
-
-  const rootReducer = combineReducers({
-    ...rootReducers
-  });
-
-  return createStore(
-    rootReducer,
-    initialState,
-    compose(applyMiddleware(...middleware), ...enhancers)
-  );
-}
